WIP ACPI display adapters driver
Currently find display adapters and attached displays but nothing more.
I would like it to publish adapters and displays in dev tree, so that
you can open and control the adapter (get firmware, boot display) and
the adapter itself would have child devices for displays (brightness ..).

However that part could use some review, and since my time is limited it
is better to make it availble for others to work on as well.