CDDA: handle CP-1252 in CD-Text
I have at least one CD which does that. There is no problem in handling
it since CP-1252 is a superset of ISO-8859-15.

Also fix the possible buffer overflow in the CD-Text handling, which
could corrupt the stack in case a track title decodes to more than 256
bytes. Now it is properly truncated on a character boundary.