Locale: fix selection of format when Default'ing
Using item->SetExpanded does not work in a BOutlineListView as the
status of the list is then out of sync. The child items are not actually
added to the visible list, and trying to select them won't work.

Using list->Expand() instead works fine, so do that.

Fixes the remaining part of #6805.