Colorcontrol: show resulting colors in the slider
Instead of showing plain red, green and blue color ramps, show what
the color would be if you moved the sliders to each point.

This is what was done on Mac OS classic and is also used in some other
tools to help you targetting a particular color.

It also limits the need for a color preview, which BColorControl
currently doesn't include, forcing apps to implement one themselves.

Fixes #3790 (the patch there didn't apply anymore, so I redid the work.
Credits for the change should go to stpere).