Web+: force download buttons to have the same width
If the text in the button is longer than what the minimal size of
BButton allows, the button exapnd. But we want the two buttons to have
the same width and stay aligned.

Fixes part of #11471.

Note: I don't understand why I need to set the explicit max size of the
button to make it bigger (I would have expected setting the minimal size
to do that), but it works.