CID1108404: missing array inintialization
BuildReportList checks that the array does not already holds the reports
before adding them. If one of the uninitialized elements happened to be
the address of one of the reports, it would not get added again,
possibly leading to items in the array being in the wrong order. Just be
safe and clear the array.

Fixes #10221, patch there not used as it cluttered the code more than
needed.