ColorControl: show the correct colors.
BColorControl was recently modified to preview the colors in the
sliders. However the computation was not correct and the colors shown
were not useful.

Fix the code so the correct colors are used. Since this can result in
the sliders having very bright colors, change the selection rings to use
black and white colors instead of just white, so they are always easily
visible.