Update BInvoker docs to clarify them
...with notes from PulkoMandy and Axel. Also added author credits.

Class documentation is moved to the appropriate method and then \sa
is used to point to the documentation so it is only documented in
one location.

Added some text about how the interaction between BInvoker and
BHandler and/or BLooper works.

BMessenger needs to be documented to understand how SetTimeout() is
suppose to work, refer to BeBook for now.