usb_midi: fix crash when unplugging devices
* When a midi device is unplugged, the driver deletes all the MIDI
channels, and unblocks all pending accesses on those. The port
structureis freed but the device kept a pointer to it.
* When a transfer is cancelled, the driver would try to notify all
callers waiting on the ports that the device is gone. But it's too late
to access the port as it was already deleted (and the callers already
unlocked).

Reset the port pointer to NULL when deleting a port, so no further
access to it is possible.

Fixes #11533.

Also remove an unused field in the usb midi device structure.