Haiku Book: compatibility: list all authors at the top of the file as per convention.
All of the other .dox files here list authors at the top of the file
in a standard (non-Doxygen) comment, so do the same here.

I've listed the file as copyright 2011, as (according to Git) that's
the last time it was functionally changed. I also added a few people
in as being authors of the file who seem to have made significant
changes to it.