src/documentation: delete the old HaikuUserGuide.txt.
Written by DarkWyrm, this user guide has been long since superseded by
the new HTML-based User Guide, which has everything this did and then
some, with the exception of the "History" section. That section
does seem like it could be useful somewhere, so I've moved it to
its own file in the "docs/misc" directory.