Interface Guidelines: migrate to docs/, use DocBookCSS.
DocBookCSS is a mostly-pure-CSS2 implementation of the DocBook standard.
Unlike DocBookXSL which relies on transforming the XML, it utilizes the
XML-styling features of modern web browsers to display the DocBook.

Its appearance still is a long way from the Haiku Book and Userguide, but
it looks (mostly) the same as the old DocBookXSL so we can stop using that.
Eventually we just need to make DocBookCSS use our styling.