interface_guidelines: add a convert.sh script that uses XMLTO.
Works on Linux, doesn't work on Haiku. It appears xsltproc tries
to fetch the DTD and XSL stylesheets (which doesn't work, for some reason
it wasn't built with HTTP support, and I can't figure out why). Even
when telling it to use the preinstalled XSL & DTDs using --catalog,
it still tries to download the files...