Tracker and Open Target Folder: fix for containers
Containers meaning directories, queries, and virtual directories

Open Target Folder was basically rewritten to work with BEntry's that are not
BDirectory's which makes it work not only for regular directories, but also for
symlinks in virtual directories and even symlinks in queries (which also didn't
work, and now does).

Update version to 1.0.1 and copyright.

Fixes the rest of #11091