app_server: Reset current view also when deleting a parent view.
The check that was in place only ensured that the current view was reset
if the current view itself got deleted. Since deleting views works by
token it is possible that a view other than the current view gets
deleted. When a parent of the current view was deleted, which also
deletes all its children, the current view pointer was not reset and
the stale pointer would still be accessed.