usb_hid: Let protocol handlers know of their closing.
On close a flag is set in the cookie of this user of a protocol handler
and the device cancels its pending transfer. This wakes up any possible
listeners. When the closed flag is set, an error code is returned from
_ReadReport() which causes the retry loop to be left. Handlers listening
on the same device which were not closed just retry the transfer.

This ensures that closing a device will cause pending control requests
to complete with a sensible error code.